requirements.txt
django==5.0.4
uvicorn==0.29.0
python-dotenv==1.0.1
psycopg2-binary==2.9.9
djangorestframework==3.15.1
markdown==3.6
django-filter==24.2
drf-spectacular==0.27.2
djangorestframework-simplejwt==5.3.1
requests==2.31.0
django-colorfield==0.11.0